Refactor drawer APIs to shared auth and tenant access helper
