+++ functions/api/drawer-assignments/create.ts
@@ -1,4 +1,5 @@
 import { neon } from "@neondatabase/serverless";
+import { canManageTenantSettings, getTenantActor, requireSessionActor } from "../../_shared/auth";
 
 const json = (data: any, status = 200) =>
   new Response(JSON.stringify(data), {
@@ -6,46 +7,10 @@ const json = (data: any, status = 200) =>
     headers: { "content-type": "application/json", "cache-control": "no-store" },
   });
 
-function readCookie(header: string, name: string): string | null {
-  if (!header) return null;
-  for (const part of header.split(/; */)) {
-    const [k, ...rest] = part.split("=");
-    if (k === name) return decodeURIComponent(rest.join("="));
-  }
-  return null;
-}
-
-async function verifyJwt(token: string, secret: string): Promise<Record<string, any>> {
-  const enc = new TextEncoder();
-  const [h, p, s] = token.split(".");
-  if (!h || !p || !s) throw new Error("bad_token");
-
-  const base64urlToBytes = (str: string) => {
-    const pad = "=".repeat((4 - (str.length % 4)) % 4);
-    const b64 = (str + pad).replace(/-/g, "+").replace(/_/g, "/");
-    const bin = atob(b64);
-    return Uint8Array.from(bin, (c) => c.charCodeAt(0));
-  };
-
-  const data = `${h}.${p}`;
-  const key = await crypto.subtle.importKey("raw", enc.encode(secret), { name: "HMAC", hash: "SHA-256" }, false, ["verify"]);
-  const ok = await crypto.subtle.verify("HMAC", key, base64urlToBytes(s), enc.encode(data));
-  if (!ok) throw new Error("bad_sig");
-
-  const payload = JSON.parse(new TextDecoder().decode(base64urlToBytes(p)));
-  if ((payload as any)?.exp && Date.now() / 1000 > (payload as any).exp) throw new Error("expired");
-  return payload as Record<string, any>;
-}
-
 export const onRequestPost: PagesFunction = async ({ request, env }) => {
   try {
-    const cookieHeader = request.headers.get("cookie") || "";
-    const token = readCookie(cookieHeader, "__Host-rp_session");
-    if (!token) return json({ ok: false, error: "no_cookie" }, 401);
-
-    const payload = await verifyJwt(token, String(env.JWT_SECRET));
-    const actor_user_id = String((payload as any).sub || "");
-    if (!actor_user_id) return json({ ok: false, error: "bad_token" }, 401);
+    const auth = await requireSessionActor(request, env, json);
+    if ("error" in auth) return auth.error;
 
     const tenant_id = request.headers.get("x-tenant-id");
     if (!tenant_id) return json({ ok: false, error: "missing_tenant" }, 400);
@@ -66,17 +31,11 @@ export const onRequestPost: PagesFunction = async ({ request, env }) => {
     }
 
     const sql = neon(String(env.DATABASE_URL));
-    const actor = await sql<{ role: string; active: boolean; can_settings: boolean }[]>`
-      SELECT m.role, m.active, COALESCE(p.can_settings, false) AS can_settings
-      FROM app.memberships m
-      LEFT JOIN app.permissions p ON p.user_id = m.user_id
-      WHERE m.tenant_id = ${tenant_id} AND m.user_id = ${actor_user_id}
-      LIMIT 1
-    `;
+    const actor = await getTenantActor(sql, tenant_id, auth.actor_user_id);
 
-    if (actor.length === 0 || actor[0].active === false) return json({ ok: false, error: "forbidden" }, 403);
-    const canManage = actor[0].role === "owner" || actor[0].role === "admin" || actor[0].role === "manager" || !!actor[0].can_settings;
-    if (!canManage) return json({ ok: false, error: "forbidden" }, 403);
+    if (!actor || actor.active === false || !canManageTenantSettings(actor)) {
+      return json({ ok: false, error: "forbidden" }, 403);
+    }
 
     const rows = await sql/*sql*/`
       INSERT INTO app.drawer_assignments (
@@ -90,7 +49,7 @@ export const onRequestPost: PagesFunction = async ({ request, env }) => {
         ${ends_at}::timestamptz,
         ${status},
         ${notes},
-        ${actor_user_id}::uuid
+        ${auth.actor_user_id}::uuid
       )
       RETURNING assignment_id, drawer_id, user_id, business_date, starts_at, ends_at, status, notes
     `;
